+++ web/wp-content/plugins/webp-converter-for-media/webp-converter-for-media.php
@@ -2,8 +2,8 @@
 
 /**
  * Plugin Name: Converter for Media
- * Description: Speed up your website by using our WebP & AVIF Converter (formerly WebP Converter for Media). Serve WebP and AVIF images instead of standard formats JPEG, PNG and GIF now!
- * Version: 6.3.0
+ * Description: Speed up your website by using our WebP & AVIF Converter. Optimize images and serve WebP and AVIF images instead of standard formats!
+ * Version: 6.3.2
  * Author: matt plugins
  * Author URI: https://url.mattplugins.com/converter-plugin-author-link
  * Text Domain: webp-converter-for-media
@@ -17,5 +17,5 @@ if ( ! defined( 'ABSPATH' ) ) {
 require_once __DIR__ . '/vendor/autoload.php';
 
 new WebpConverter\WebpConverter(
-	new WebpConverter\PluginInfo( __FILE__, '6.3.0' )
+	new WebpConverter\PluginInfo( __FILE__, '6.3.2' )
 );