+++ web/wp-content/plugins/wp-cloudflare-page-cache/bootstrap.php
@@ -21,7 +21,7 @@ if ( ! class_exists( 'SW_CLOUDFLARE_PAGECACHE' ) ) {
 	define( 'SWCFPC_AUTH_MODE_API_TOKEN', 1 );
 	define( 'SWCFPC_LOGS_STANDARD_VERBOSITY', 1 );
 	define( 'SWCFPC_LOGS_HIGH_VERBOSITY', 2 );
-	define( 'SWCFPC_VERSION', '5.1.5' );
+	define( 'SWCFPC_VERSION', '5.2.0' );
 	if ( ! defined( 'SPC_METRICS_DIR' ) ) {
 		$home_url_parts = parse_url( home_url() );
 		define( 'SPC_METRICS_DIR', WP_CONTENT_DIR . "/wp-cloudflare-super-page-cache/{$home_url_parts['host']}/metrics" );